+++ react-app/src/components/Gallery/Empty.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Redirect } from "react-router-dom";
-import { Helmet } from "react-helmet";
+import { Helmet } from "react-helmet-async";
 import { BsFillHouseFill } from "react-icons/bs";
 
 import Root from "./Navigation";