+++ react-app/src/lib/config.js
@@ -3,32 +3,32 @@
  *
  * This can be overridden with the instance metadata `cdn`.
  */
-var PHOTO_ROOT_URL = process.env.REACT_APP_PHOTO_ROOT_URL || "/";
+const env = import.meta.env;
+var PHOTO_ROOT_URL = env.VITE_PHOTO_ROOT_URL || "/";
 /**
  * The default language to choose if the user has not selected one.
  */
-const DEFAULT_LANGUAGE = process.env.REACT_APP_DEFAULT_LANGUAGE || "en";
+const DEFAULT_LANGUAGE = env.VITE_DEFAULT_LANGUAGE || "en";
 /**
  * see "themes.css"
  */
-const DEFAULT_THEME = process.env.REACT_APP_THEME || "blue";
+const DEFAULT_THEME = env.VITE_THEME || "blue";
 /**
  * The default gallery to redirect to when accessing the galleries end-point.
  */
-const DEFAULT_GALLERY = process.env.REACT_APP_DEFAULT_GALLERY || undefined;
+const DEFAULT_GALLERY = env.VITE_DEFAULT_GALLERY || undefined;
 /**
  * The view to redirect to from the gallery root end-point, to the last element of the respective view.
  *
  * One of "year", "month", "day", "photo"
  */
-const INITIAL_GALLERY_VIEW =
-  process.env.REACT_APP_INITIAL_GALLERY_VIEW || "month";
+const INITIAL_GALLERY_VIEW = env.VITE_INITIAL_GALLERY_VIEW || "month";
 /**
  * The first day of the week for the calendar grid of the year view.
  *
  * 1 = Monday, 0 = Sunday
  */
-const FIRST_WEEKDAY = process.env.REACT_APP_FIRST_WEEKDAY || 1;
+const FIRST_WEEKDAY = env.VITE_FIRST_WEEKDAY || 1;
 
 let lang = DEFAULT_LANGUAGE;